Rename path time floe by time billingx and change routes.
