+++ docs/containers/Routes.js
@@ -105,10 +105,10 @@ const Routes = () =>
         <Route
           getComponent={(_nextState, callback) => {
             require.ensure([], require => {
-              callback(null, require('../examples/timeflow'))
-            }, 'example-timeflow')
+              callback(null, require('../examples/timebillingx'))
+            }, 'example-timebillingx')
           }}
-          path="timeflow"
+          path="timebillingx"
         />
       </Route>
     </Route>
